Move context tests bak into context_test.rb and add improved environemnt tests
